Texel: Fixed portability problems.
git-svn-id: file:///home/petero/vc/svnrepos/texel/trunk@39 4abc0a36-5a1e-4bec-8e7d-414f7125f69d

+++ texellib/src/game.cpp
@@ -247,7 +247,7 @@ Game::getPosHistory(std::vector<std::string> ret) {
     }
     ret.push_back(moves); // Store move list string
 
-    int numUndo = moveList.size() - currentMove;
+    int numUndo = (int)moveList.size() - currentMove;
     ret.push_back(num2Str(numUndo));
 }
